+++ Craftsman/Helpers/ApiScaffolding.cs
@@ -135,10 +135,10 @@
                 AddBusCommand.AddBus(template.Bus, srcDirectory, projectBaseName, solutionDirectory, fileSystem);
 
             if (template.Consumers.Count > 0)
-                RegisterConsumerCommand.AddConsumers(template.Consumers, projectBaseName, srcDirectory);
+                AddConsumerCommand.AddConsumers(template.Consumers, projectBaseName, srcDirectory);
 
             if (template.Producers.Count > 0)
-                RegisterProducerCommand.AddProducers(template.Producers, projectBaseName, srcDirectory);
+                AddProducerCommand.AddProducers(template.Producers, projectBaseName, srcDirectory);
 
             if (template.Messages.Count > 0)
                 AddMessageCommand.AddMessages(solutionDirectory, fileSystem, template.Messages);